+++ project/src/models/events/ticket_event.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 from ....setup import db
 from enum import Enum
 from ...utils.time import TimeUtil
-from ..models.user import User
-from ..models.ticket import Ticket
+from ..user import User
+# from ..ticket import Ticket
 from typing import List
-from ..models.course import Course
+# from ..course import Course
 
 
 class EventType(Enum):
@@ -30,7 +30,7 @@ class EventType(Enum):
     COMMENTED = 6
 
 
-class TicketEvent(db.model):
+class TicketEvent(db.Model):
     """
     The event happened on ticket.\n
     Fields:\n
@@ -52,7 +52,7 @@ class TicketEvent(db.model):
     is_private = db.Column(db.Boolean, nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'),
                         nullable=False)
-    timestamp = db.Column(db.Datetime, nullable=False,
+    timestamp = db.Column(db.DateTime, nullable=False,
                           default=TimeUtil.get_current_time())
 
     # Getter Methods
